+++ server/src/dbnet/tls.rs
@@ -30,7 +30,11 @@ use crate::dbnet::tcp::Connection;
 use crate::dbnet::BaseListener;
 use crate::dbnet::Terminator;
 use libsky::TResult;
+use openssl::pkey::PKey;
+use openssl::rsa::Rsa;
 use openssl::ssl::{Ssl, SslAcceptor, SslFiletype, SslMethod};
+use std::fs;
+use std::io::Error as IoError;
 use std::pin::Pin;
 use tokio::net::TcpStream;
 use tokio::time::{self, Duration};
@@ -48,12 +52,36 @@ impl SslListener {
         key_file: String,
         chain_file: String,
         base: BaseListener,
+        tls_passfile: Option<String>,
     ) -> TResult<Self> {
-        let mut acceptor = SslAcceptor::mozilla_intermediate(SslMethod::tls())?;
-        acceptor.set_private_key_file(key_file, SslFiletype::PEM)?;
-        acceptor.set_certificate_chain_file(chain_file)?;
-        let acceptor = acceptor.build();
-        Ok(SslListener { base, acceptor })
+        let mut acceptor_builder = SslAcceptor::mozilla_intermediate(SslMethod::tls())?;
+        // cert is the same for both
+        acceptor_builder.set_certificate_chain_file(chain_file)?;
+        if let Some(tls_passfile) = tls_passfile {
+            // first read in the private key
+            let tls_private_key = fs::read(key_file).map_err(|e: IoError| {
+                format!("Failed to read TLS private key file with error: {}", e)
+            })?;
+            // read the passphrase because the passphrase file stream was provided
+            let tls_keyfile_stream = fs::read(tls_passfile).map_err(|e: IoError| {
+                format!(
+                    "Failed to read TLS private key passphrase file with error: {}",
+                    e
+                )
+            })?;
+            // decrypt the private key
+            let pkey = Rsa::private_key_from_pem_passphrase(&tls_private_key, &tls_keyfile_stream)?;
+            let pkey = PKey::from_rsa(pkey)?;
+            // set the private key for the acceptor
+            acceptor_builder.set_private_key(&pkey)?;
+        } else {
+            // no passphrase, needs interactive
+            acceptor_builder.set_private_key_file(key_file, SslFiletype::PEM)?;
+        }
+        Ok(SslListener {
+            base,
+            acceptor: acceptor_builder.build(),
+        })
     }
     async fn accept(&mut self) -> TResult<SslStream<TcpStream>> {
         let mut backoff = 1;